Use conditional imports so that dart:io is not imported from the main test.dart import.
